# api/openapi.yaml
openapi: 3.0.3
info:
  title: GoSheet API
  description: |
    HTTP API for GoSheet spreadsheet application.

    ## Authentication

    All `/api/*` endpoints require a bearer token. Two token types are in use:

    **Bootstrap token** (`BootstrapAuth`): Generated by the Go server at startup using a 32-byte
    CSPRNG (base64url encoded). Written to fd 3 alongside the port number. Electron injects it into
    the renderer, which sends it as `Authorization: Bearer <token>` on every request. Has full
    access to all endpoints. In `NODE_ENV=test` mode the server skips token generation and auth is
    disabled entirely so Playwright tests run without modification.

    **Agent token** (`AgentAuth`): Scoped token (`ro` or `rw`) issued by `POST /api/agent/token`
    in exchange for a valid bootstrap token. Used exclusively by AI agents. Agent tokens are
    rejected on `/api/file/*` endpoints regardless of scope. Read-only (`ro`) tokens are rejected
    on any mutating agent endpoint.

    The bootstrap endpoint (`GET /api/agent/bootstrap`) also accepts the agent token as a
    `?token=` query parameter for copy-paste UX when the user hands a URL to an agent.
  version: 1.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Relative to host (localhost ephemeral port)

# All /api/* endpoints require auth. Static file routes are unauthenticated.
# In NODE_ENV=test the server disables auth entirely (bootstrapToken == "").
security:
  - BootstrapAuth: []
components:
  securitySchemes:
    BootstrapAuth:
      type: http
      scheme: bearer
      description: |
        Bootstrap token generated at server startup. Injected into the Electron renderer via URL
        query param and forwarded as `Authorization: Bearer <token>` on every API request.
        Has full access to all endpoints including `/api/file/*`.
    AgentAuth:
      type: http
      scheme: bearer
      description: |
        Scoped agent token issued by `POST /api/agent/token`. Used by AI agents.
        Scope `rw` allows reads and writes; scope `ro` allows reads only.
        Rejected on `/api/file/*` endpoints regardless of scope.
...
paths:
  /api/cell/value:
    get:
      operationId: getCellValue
      summary: Get cell computed value
      parameters:
        - name: row
          in: query
          required: true
          schema:
            type: integer
        - name: col
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cell value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellValueResponse'
        '400':
          description: Bad request (invalid or missing row/col)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cell/raw:
    get:
      operationId: getCellRaw
      summary: Get cell raw value (formula or text)
      parameters:
        - name: row
          in: query
          required: true
          schema:
            type: integer
        - name: col
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cell raw value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellRawResponse'
        '400':
          description: Bad request (invalid or missing row/col)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cell/set:
    post:
      operationId: setCellValue
      summary: Set cell value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCellRequest'
      responses:
        '200':
          description: Cell updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetCellResponse'
        '400':
          description: Bad request (invalid body, missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent token with ro scope, or agent token on file endpoint)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. circular ref, invalid formula)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cell/ref:
    get:
      operationId: getCellRef
      summary: Get cell reference (e.g. A1)
      parameters:
        - name: row
          in: query
          required: true
          schema:
            type: integer
        - name: col
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cell reference string
          content:
            application/json:
              schema:
                type: string
                example: '"A1"'
        '400':
          description: Bad request (invalid or missing row/col)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/merges:
    get:
      operationId: getMerges
      summary: Get all merge regions
      responses:
        '200':
          description: Merge regions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/merge:
    post:
      operationId: setMerge
      summary: Create merge region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMergeRequest'
      responses:
        '200':
          description: Merge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetMergeResponse'
        '400':
          description: Bad request (overlap, invalid bounds)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent token with ro scope)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/unmerge:
    post:
      operationId: unmerge
      summary: Remove merge region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnmergeRequest'
      responses:
        '200':
          description: Merge removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnmergeResponse'
        '400':
          description: Bad request (anchor not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cell/style:
    post:
      operationId: applyCellStyle
      summary: Apply style to a single cell
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyCellStyleRequest'
      responses:
        '200':
          description: Style applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyStyleResponse'
        '400':
          description: Bad request (invalid row/col/styleId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent token with ro scope)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/range/style:
    post:
      operationId: applyRangeStyle
      summary: Apply style to a range of cells
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyRangeStyleRequest'
      responses:
        '200':
          description: Style applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyStyleResponse'
        '400':
          description: Bad request (invalid range/styleId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent token with ro scope)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/format/cleanup:
    post:
      operationId: formatCleanup
      summary: Remove style from empty cells and delete cells with no value and no style
      responses:
        '200':
          description: Cleanup completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyStyleResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method not allowed (POST required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cells/all:
    get:
      operationId: getAllCells
      summary: Get all non-empty cells
      responses:
        '200':
          description: Array of cell data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllCellsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/save:
    post:
      operationId: saveFile
      summary: Save spreadsheet to file
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: File saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveLoadResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. file write error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/load:
    post:
      operationId: loadFile
      summary: Load spreadsheet from file
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: File loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveLoadResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. file not found, read error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/new:
    post:
      operationId: newFile
      summary: Create new empty spreadsheet
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      responses:
        '200':
          description: New file created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/status:
    get:
      operationId: getFileStatus
      summary: Get file status
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      responses:
        '200':
          description: File status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/download:
    get:
      operationId: downloadFile
      summary: Download spreadsheet as binary file
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      responses:
        '200':
          description: Binary .gosheet file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/upload:
    post:
      operationId: uploadFile
      summary: Upload spreadsheet from binary data
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request (invalid body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/csv/preview:
    post:
      operationId: previewCsv
      summary: Preview CSV file before import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: CSV preview data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CSVPreviewResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. file read error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/csv/import:
    post:
      operationId: importCsv
      summary: Import CSV file into spreadsheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CSVImportResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. parse error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/formula/shift:
    post:
      operationId: shiftFormula
      summary: Shift cell/range references in a formula string by a row/col offset
      description: Stateless pure transformation — no spreadsheet state is read or written. Relative refs that shift out of bounds become #REF!.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftFormulaRequest'
      responses:
        '200':
          description: Shifted formula string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftFormulaResponse'
        '400':
          description: Bad request (invalid body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method not allowed (POST required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/csv/export:
    post:
      operationId: exportCsv
      summary: Export spreadsheet to CSV file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: Export result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CSVExportResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. file write error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Agent endpoints (Epic 20) ──────────────────────────────────────────────
  /api/agent/token:
    post:
      operationId: issueAgentToken
      summary: Issue a scoped agent token
      description: |
        Exchanges the bootstrap token for a scoped agent token. Only one agent session
        may be active per file at a time — a second call returns 409 while a session is active.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTokenRequest'
      responses:
        '200':
          description: Agent token issued, session opened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTokenResponse'
        '400':
          description: Bad request (invalid scope)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (missing or invalid bootstrap token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict — an agent session is already active for this file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/agent/workbook:
    get:
      operationId: getAgentWorkbook
      summary: Get workbook summary (file path, dimensions, cell count)
      description: Accessible with ro or rw scope.
      security:
        - AgentAuth: []
      responses:
        '200':
          description: Workbook summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentWorkbookResponse'
        '401':
          description: Unauthorized (missing or invalid agent token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...